backend/requirements.txt
flask
flask-cors
python-dotenv
google-genai
numpy
scipy
scikit-learn
umap-learn
hdbscan
youtube-transcript-api
requests
